RegisterPage, UserPage and ButtonComponent edit
