+++ package.json
@@ -16,6 +16,7 @@
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-hook-form": "^7.59.0",
+    "react-hot-toast": "^2.6.0",
     "react-router": "^7.6.3",
     "react-router-dom": "^7.6.2",
     "reactjs-popup": "^2.0.6"
