+++ src/components/buttonComponent/ButtonComponent.jsx
@@ -1,6 +1,6 @@
 import './ButtonComponent.scss';
 
-function ButtonComponent( {classNames, buttonText, noteIcon, downloadIcon, editIcon, buttonFunction }) {
+function ButtonComponent( {classNames, buttonText, noteIcon, downloadIcon, editIcon, reviewIcon, buttonFunction }) {
     console.log(classNames);
     return (
         <button className={`btn ${classNames}`} onClick={buttonFunction}>
@@ -8,6 +8,7 @@ function ButtonComponent( {classNames, buttonText, noteIcon, downloadIcon, editI
             { noteIcon && <> <i className="fa-solid fa-music"></i></> }
             { downloadIcon && <> <i className="fa-solid fa-download"></i></> }
             { editIcon && <> <i className="fa-solid fa-gear"></i></> }
+            {reviewIcon && <> <i className="fa-solid fa-pen"></i></> }
         </button>
     );
 }
