+++ src/helpers/dateFormatter.js
@@ -1,0 +1,13 @@
+function dateFormatter(date) {
+    const options = {
+        day: 'numeric',
+        month: 'long',
+        year: 'numeric'
+    }
+
+    const formattedDate = new Date(date).toLocaleDateString('nl-NL', options);
+
+    return formattedDate;
+}
+
+export default dateFormatter;
